GAV-125: Mostrar el costo de cancelación (#33)
* Add cancellation cost field

* Update tests

* Update docker compose file

* Add cancellation cost column to database seeding script

* Add cancellation cost to the /assistance endpoint

* Improve validations and messages

* Fix test class name

* Remove unused variable

+++ src/main/kotlin/api/dtos/OrderUpdateRequestDTO.kt
@@ -5,6 +5,6 @@ import java.util.*
 class OrderUpdateRequestDTO(
     val orderId: UUID?,
     val status: String?,
-    val kmTraveled: Int?,
+    val kmTraveled: Double?,
     val password: String?
 )
